+++ src/components/Details.js
@@ -323,13 +323,15 @@ const Details = ({
 
       if (previousTransactions.length > 0) {
         // 이번 달 데이터 삭제
-        await axios.delete(`${BASE_URL}/api/deleteAll`, {
-          params: {
-            userId: user.id,
-            monthId: monthData.id,
-            type: categoryTitle,
-          },
-        });
+        if (monthData.id) {
+          await axios.delete(`${BASE_URL}/api/deleteAll`, {
+            params: {
+              userId: user.id,
+              monthId: monthData.id,
+              type: categoryTitle,
+            },
+          });
+        }
 
         // 이전 달 데이터 추가하기
         const promises = previousTransactions.map(async (transaction) => {
